+++ src/test/java/jewellery/inventory/unit/service/UserInOrganizationServiceTest.java
@@ -3,9 +3,7 @@ package jewellery.inventory.unit.service;
 import static jewellery.inventory.helper.OrganizationTestHelper.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 import java.util.*;
 import jewellery.inventory.dto.response.*;
@@ -54,9 +52,6 @@ class UserInOrganizationServiceTest {
   void getUsersInOrganizationSuccessfully() {
     when(organizationService.getOrganization(organizationWithUserAllPermission.getId()))
         .thenReturn(organizationWithUserAllPermission);
-    ;
-    when(authService.getCurrentUser()).thenReturn(executorResponseDto);
-    when(userService.getUser(user.getId())).thenReturn(user);
 
     when(organizationMapper.toOrganizationMembersResponseDto(organizationWithUserAllPermission))
         .thenReturn(new OrganizationMembersResponseDto());
@@ -70,9 +65,6 @@ class UserInOrganizationServiceTest {
 
   @Test
   void updateUserPermissionsInOrganizationSuccessfully() {
-    when(organizationService.getOrganization(organizationWithUserAllPermission.getId()))
-        .thenReturn(organizationWithUserAllPermission);
-
     when(userInOrganizationRepository.findByUserIdAndOrganizationId(
             organizationWithUserAllPermission.getUsersInOrganization().get(0).getUser().getId(),
             organizationWithUserAllPermission.getId()))
@@ -96,8 +88,9 @@ class UserInOrganizationServiceTest {
         .thenReturn(organizationWithUserAllPermission);
 
     doThrow(MissingOrganizationPermissionException.class)
-            .when(organizationService)
-            .validateCurrentUserPermission(organizationWithUserAllPermission, OrganizationPermission.MANAGE_USERS);
+        .when(organizationService)
+        .validateCurrentUserPermission(
+            organizationWithUserAllPermission, OrganizationPermission.MANAGE_USERS);
 
     assertThrows(
         MissingOrganizationPermissionException.class,
@@ -110,8 +103,9 @@ class UserInOrganizationServiceTest {
   void getUsersInOrganizationThrowsExceptionWhenUserIsNotPartOfOrganizationException() {
     when(organizationService.getOrganization(organizationWithUserAllPermission.getId()))
         .thenReturn(organizationWithUserAllPermission);
-    when(authService.getCurrentUser()).thenReturn(executorResponseDto);
-    when(userService.getUser(any(UUID.class))).thenReturn(new User());
+    doThrow(UserIsNotPartOfOrganizationException.class)
+        .when(organizationService)
+        .validateUserInOrganization(organizationWithUserAllPermission);
 
     assertThrows(
         UserIsNotPartOfOrganizationException.class,
